Working output, now working on motors

+++ adc.h
@@ -8,4 +8,4 @@
 
 void initAdc();
 
-int readAdc(int val);
+int readAdc(float val);
